[SPARK-51310][SQL] Resolve the type of default string producing expressions
### What changes were proposed in this pull request?
Add casts to `DefaultStringProducingExpression` to have the correct type based on the default string type.

Currently, we can only test this on views, until we implement schema level collations when we will be able to test the behavior on newly created tables as well.

### Why are the changes needed?
These expressions should return the default string type, which could be different in some DDL commands.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Added unit tests for view creation.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #50053 from stefankandic/resolveDefaultStringExprs.

Authored-by: Stefan Kandic <stefan.kandic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>