[MINOR][TESTS][CONNECT] Fix the teardown function of test_connect_function.py
### What changes were proposed in this pull request?
It is a small fix to a typo in teardown function of a test where instead of calling teardown of the parent object we are calling setup.

### Why are the changes needed?
This typo needs to be fixed

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
This is already covered by the changed test.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #50120 from sfc-gh-dyadav/patch-1.

Authored-by: Deepak Yadav <deepak.yadav@snowflake.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 737e2b407b5d166374e0d8b43fac9eaf9ffb8adb)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>