[SPARK-51373][SS] Removing extra copy for column family prefix from 'ReplyChangelog'
### What changes were proposed in this pull request?

There is currently an extra copy when column families are enabled to first remove the column family prefix, and then one to add this column family prefix back when replaying the changelog. Because we don't do anything with this prefix or the raw bytes, we don't need to add and remove and can instead just passthrough.

### Why are the changes needed?

This change is needed for performance reasons for the TransformWithState operator - extra copies incur performance hit, and this change removes these extra copies.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing unit tests are sufficient, as we are not adding any new functionality.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50119 from ericm-db/changelog-copy.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
(cherry picked from commit 050226f740332b2099301660944d8db5285418e8)
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>