[SPARK-51396][SQL] RuntimeConfig.getOption shouldn't use exceptions for control flow
### What changes were proposed in this pull request?

RuntimeConfig.getOption is defined as

```scala
  /** inheritdoc */
  def getOption(key: String): Option[String] = {
    try Option(get(key)) catch {
      case _: NoSuchElementException => None
    }
  }
```

This uses exceptions for control flow, which can be a performance problem.

This PR updates this code to use a new package-private `getConfStringOption` helper in order to avoid the exceptions.

### Why are the changes needed?

Address a perf. problem if `getOption` for non-existing configurations is called at a high frequency. It's usually not a good practice to perform configuration lookups at a high rate (e.g in a hot method or loop), but if someone _does_ do that then the exceptions make the perf. issue worse.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing unit tests in `RuntimeConfigSuite`.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50167 from JoshRosen/no-exceptions-for-control-flow-in-runtimeconf-getoption.

Authored-by: Josh Rosen <joshrosen@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>