Revert "[SPARK-51396][SQL] RuntimeConfig.getOption shouldn't use exceptions for control flow"
### What changes were proposed in this pull request?

This reverts commit 6bf6c464605c0f9ac2eedf404be0727a75630624 / https://github.com/apache/spark/pull/50167.

That PR introduced a subtle bug which could lead to NullPointerExceptions in the `SET` command if any `ConfigEntry`'s string form was `null`, which can happen when reading an `OptionalConfigEntry` where no value is set.

I have some ideas for a fix-forward, but for now let's revert so as not to leave things in a broken state.

### Why are the changes needed?

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Clean revert, n/a.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50200 from JoshRosen/revert-SPARK-51396.

Authored-by: Josh Rosen <joshrosen@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>