[SPARK-51443][SS] Fix singleVariantColumn in DSv2 and readStream
### What changes were proposed in this pull request?

The current JSON `singleVariantColumn` mode doesn't work in DSv2 and `spark.readStream`. This PR fixes the two cases:
- DSv1 calls `JsonFileFormat.inferSchema`, which calls `JsonDataSource.inferSchema`; DSv2 calls `JsonDataSource.inferSchema`. The previous `singleVariantColumn` code was in `JsonFileFormat.inferSchema`, and is now moved into `JsonDataSource.inferSchema`, so that both cases can be covered.
- `spark.readStream` requires that there must be a user-specified schema. `singleVariantColumn` plays the same row as a user-specified schema, but the check would fail.

It also includes a small refactor that moves the option name definition `singleVariantColumn` from `JSONOptions` to `DataSourceOptions`. It will be a common option name shared by multiple data sources (e.g., CSV) when we add the implementation in the future.

### Why are the changes needed?

It is a bug fix that improves the usability of variant.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unit test. A test previously in `VariantSuite` is moved to `JsonSuite`, so that we can test the read behavior in both `JsonV1Suite` and `JsonV2Suite`. The test is also extended to include `spark.readStream`.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50217 from chenhao-db/fix_json_singleVariantColumn.

Authored-by: Chenhao Li <chenhao.li@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit fa6f18fdb8c7062f80cb8fac464a0e604fd3c525)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>