[SPARK-51542][UI] Add a scroll-button for addressing top and bottom
### What changes were proposed in this pull request?
This PR adds buttons for addressing the top and bottom of every page

### Why are the changes needed?
This is a UI UX improvement.

Currently, most of our pages cannot be shown on a single screen and the nav bar on top of them is not fixed to the screen, a.k.a, it disappears when we scroll down.

This button meets the need for quickly locating the nav bar and/or something else on our UI.

### Does this PR introduce _any_ user-facing change?
yes, an UI only change

### How was this patch tested?
A new jest test for js added

![image](https://github.com/user-attachments/assets/26b1f24e-3ced-4992-8523-e1dddd4ec1b6)

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #50305 from yaooqinn/SPARK-51542.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: yangjie01 <yangjie01@baidu.com>