[SPARK-51559][SQL] Make max broadcast table size configurable
### What changes were proposed in this pull request?

Currently the maximum size for a table to be broadcast in `BroadcastExchangeExec` is hard-coded to 8GB, and it seems there is no particular reason why this value was picked initially. In certain scenarios, we may want broadcast hash join to be triggered even with larger table sizes, since otherwise it will fallback to expensive shuffle join.

### Why are the changes needed?

This PR introduces a new configuration: `spark.sql.maxBroadcastTableSize`, which by default still uses 8GB, so no behavior change expected. This would allow users to update the config when necessary to be more flexible.

### Does this PR introduce _any_ user-facing change?

No behavior change. Only a new config is added.

### How was this patch tested?

We tested out this internally and it seems to be working well.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50327 from sunchao/SPARK-51559.

Authored-by: Chao Sun <chao@openai.com>
Signed-off-by: Chao Sun <chao@openai.com>