[SPARK-51584][SQL] Add rule that pushes Project through Offset and Suite that tests it
### What changes were proposed in this pull request?

Add a new rule PushProjectionThroughOffset analogous to PushProjectionThroughLimit, that pushes projection nodes beneath Offset nodes.

### Why are the changes needed?

Currently, if a Project node is above a Offset and Limit node, the Project will get pushed through the Limit node, but not the Offset node, and then the Offset and LocalLimit nodes won't get swapped, eliminating the optimization LocalLimit nets in this case.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Created PushProjectionThroughOffsetSuite to test this new rule.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50326 from Pajaraja/pavle-martinovic_data/ProjectionPushDownThroughLimit.

Authored-by: pavle-martinovic_data <pavle.martinovic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>