[SPARK-51594][SQL] Use empty schema when saving a view which is not Hive compatible
### What changes were proposed in this pull request?

This is a long-standing issue. Spark always tries to save views in the Hive-compatible way, and only set the schema to empty if the save operation fails. However, for certain Hive compatibility issues, the save operation works but the following read operations will fail.

This PR fixes this issue by setting view schema to empty if it's not Hive compatible.

### Why are the changes needed?

to not create malformed views that no one can read.

### Does this PR introduce _any_ user-facing change?

Yes, the malformed view will be saved in non hive compatible way so that Spark can read it.

### How was this patch tested?

updated test case

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #50367 from cloud-fan/view.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>