[SPARK-51625][SQL] Command in CTE relations should trigger inline
### What changes were proposed in this pull request?

This PR fixes a small issue in the rule `CTESubstitution`. The intention is to fall back to CTE inline if there are commands in the plan. However, the code simply does `plan.collect` which misses the fact that `UnresolvedWith` does not include the CTE relations in its children.

To fix it, we should do a manually recursion to handle the CTE relations.

### Why are the changes needed?

A potential bug as the code does not work as expected.

### Does this PR introduce _any_ user-facing change?

No, there is no user-facing API that can construct such a CTE plan.

### How was this patch tested?

a new test

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #50425 from cloud-fan/cte.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>