[SPARK-51252][SS][TESTS] Make tests more consistent for StateStoreInstanceMetricSuite
### What changes were proposed in this pull request?

SPARK-51252

Fix for flaky tests described in #50030.

This PR relaxes the version check in the tests and processes more data in queries to force more batches.

I also reduced the number of shuffle partitions for one of the join queries because there used to be 5 * 4 state stores that would need a snapshot to upload in a short time span. Now there's only 3 * 4 stores to deal with.

### Why are the changes needed?

The tests previously introduced in #50030 are quite flaky depending on the environment, resulting in CI gates failing some tests in the suite.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Re-ran and verified the tests multiple times locally, but primarily using the current merge gates to see if these pass properly.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50405 from zecookiez/SPARK-51252-hdfs-flaky-fix.

Authored-by: Zeyu Chen <zycm03@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>