[SPARK-51657] UTF8_BINARY default table collation shown by default in Desc As JSON (v1)
### What changes were proposed in this pull request?

Display `UTF8_BINARY` default table collation by default in Desc As JSON for v1 table.

Note this only affects the `desc as json` output and not `desc table extended` -- desc table does NOT always show the top-level table collation (if omitted in the`desc table` output, it is `UTF8_BINARY`)

### Why are the changes needed?

If the CatalogTable `collation` property is None, then it is by default `UTF8_BINARY`. This update is for completeness for the `desc as json` output.

### Does this PR introduce _any_ user-facing change?

Yes, affects the JSON result

### How was this patch tested?

Added tests in `v1/DescribeTableSuite.scala`

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50451 from asl3/asl3/descasjson-defaultcollation.

Authored-by: Amanda Liu <amanda.liu@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>