[SPARK-51687][SQL] Pushdown filters with TIME values to parquet
### What changes were proposed in this pull request?
In the PR, I propose to support pushdown filters over TIME columns in the parquet datasource. In particular, pushdown the following operators: `===, <=>, =!=, isNull, isNotNull, >, >=, <, <=`.

### Why are the changes needed?
The PR https://github.com/apache/spark/pull/50389 supported the TIME type in the Parquet datasource as INT64 but filters with such values are not pushed down. So, comparing to primitive types, the TIME type can suffer from worse performance in read. This PR aims to solve the issue, and achieve feature parity with other types.

### Does this PR introduce _any_ user-facing change?
No, the changes might influence on performance of the parquet datasource only.

### How was this patch tested?
By running new tests in `ParquetFilterSuite`:
```
$ build/sbt clean "test:testOnly *ParquetV1FilterSuite"
$ build/sbt clean "test:testOnly *ParquetV2FilterSuite"
```

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #50473 from MaxGekk/parquet-time-filter.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>