[SPARK-51298][SQL] Support variant in CSV scan
### What changes were proposed in this pull request?

It supports reading CSV data into the variant data type (including CSV scan and the from_csv expression). Writing variant data as CSV currently is and will remain unsupported.

If a column is specified as the variant type when reading CSV data, the corresponding comma-separated value will be converted into variant data. The CSV parser will maintain the current content type of the variant column (long, decimal, date, timestamp, boolean). When encountering a new value, it tries to parse the input as the current content type. If the parsing fails, the parser moves to the next type and continues the trial. It either successfully parses the input as a specific scalar type, or fails after trying all the types and defaults to the string type.

Similar to JSON scan, we will support the singleVariantColumn option in CSV scan. The option value is a string that specifies the name of the single output column of variant type. When this option is set, the entire line of CSV data is collected into a single variant object. The existing header option in CSV scan interacts with singleVariantColumn. If header is set to true when using singleVariantColumn, the first line of the CSV file will be treated as the header row, and the field names will be extracted from it.

### Why are the changes needed?

It allows users to ingest CSV data with variant.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50052 from chenhao-db/variant_csv_scan.

Authored-by: Chenhao Li <chenhao.li@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>