[SPARK-51079] Added constructor for ArrowSerializer for backward compatibility
### What changes were proposed in this pull request?

Previously, [PR](https://github.com/apache/spark/pull/49790) extends ArrowSerializer class to support large variable types in pandas UDF. This PR makes it more backward-compatible by adding a new constructor for cases where the flag for large variable is irrelevant (default to false).

### Why are the changes needed?
To achieve versionless in spark connect, we need to be careful about changing a client API in Scala. For example, the change of an interface like this could affect Scala UDF referring to this class (as it is part of the spark connect client package) but running with an older spark connect client.

### How was this patch tested?
N/A. It's a new constructor.

Closes #50513 from haiyangsun-db/SPARK-51079.

Authored-by: Haiyang Sun <75672016+haiyangsun-db@users.noreply.github.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>