[SPARK-51395][SQL] Refine handling of default values in procedures
### What changes were proposed in this pull request?

This PR refines handling of default values in procedures that will be released in 4.0.

### Why are the changes needed?

These changes are needed as connectors like Iceberg may not have utilities to generate SQL strings containing Spark SQL dialects. The API should be changed to allow either a DSv2 expression or a SQL string.

### Does this PR introduce _any_ user-facing change?

Yes, but the stored procedure API hasn't been released yet.

### How was this patch tested?

This PR comes with tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50197 from aokolnychyi/spark-51395.

Authored-by: Anton Okolnychyi <aokolnychyi@apache.org>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>