[SPARK-51296][SQL] Support collecting corrupt data in singleVariantColumn mode
### What changes were proposed in this pull request?

Currently, if the `singleVariantColumn` is specified, the schema will be a single variant column. It is then impossible to collect corrupt data, which requires the schema to contain a column for corrupt data. This PR enables collecting corrupt data in `singleVariantColumn` mode by relaxing the requirement that `singleVariantColumn` must not exist together with the user-specified schema. The new requirement is that, if the user-specified schema exists, it must either be a variant field, or a variant field plus a corrupt column field. Corrupt data can be collected in the latter case.

### Why are the changes needed?

It allows collecting corrupt data in `singleVariantColumn` mode.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50517 from chenhao-db/single_variant_column_corrupt_record.

Authored-by: Chenhao Li <chenhao.li@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>