[SPARK-51732][SQL] Apply `rpad` on attributes with same `ExprId` if they need to be deduplicated
### What changes were proposed in this pull request?
This PR fixes a case where `rpad` is not applied on attributes that have the same `ExprId` even though those attributes should be deduplicated.

### Why are the changes needed?
For example, consider the following query:

```
CREATE OR REPLACE TABLE t(a CHAR(50));
SELECT t1.aFROM t t1
WHERE (SELECT count(*) AS item_cnt FROM t t2 WHERE (t1.a = t2.a)) > 0
```
In the above case, `ApplyCharTypePadding` will run for subquery where `t1.a` and `t2.a` will reference the same `ExprId`, therefore we won't apply `rpad`. However, after `DeduplicateRelations` runs for outer query, `t1.a` and `t2.a` will get different `ExprIds` and would therefore need `rpad`. However, this doesn't happen because `ApplyCharTypePadding` for outer query does not recurse into the subquery.

On the other hand, for a query:

```
SELECT t1.a
FROM t t1, t t2
WHERE t1.a = t2.a
```
`ApplyCharTypePadding` will correctly add `rpad` to both `t1.a` and `t2.a` because attributes will first be deduplicated.

In particular, this fixes a code-path when `readSideCharPadding` is off and `LEGACY_NO_CHAR_PADDING_IN_PREDICATE` is also false

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #50527 from mihailotim-db/mihailotim-db/apply_char_type_padding_subqueries.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>