[SPARK-51673][SQL] Apply default collation to alter view query
### What changes were proposed in this pull request?
Fixed the application of default collation in `ALTER VIEW` queries. For example, if a view has a default collation and we execute `ALTER VIEW v AS SELECT 'a' AS c1`, the default collation was not being applied to the literal `a`.

### Why are the changes needed?
Bug fix.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Tests added to `DefaultCollationTestSuite.scala`

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #50468 from ilicmarkodb/fix_alter_view.

Authored-by: ilicmarkodb <marko.ilic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>