[SPARK-51751][SQL] Fix multiple rCTEs for one WITH statement that reference each other
### What changes were proposed in this pull request?

Add a check to UnionLoopExec that the UnionLoopRef in its subtree is referring to the correct query. This is important since if we call another rCTE from it, it will also have a UnionLoopRef which needs to not be replaced with the reference.

### Why are the changes needed?

Multiple rCTEs within one WITH statement ended in an infinite recursion.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

New golden file test added.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #50544 from Pajaraja/pavle-martinovic_data/multiplerctesfix.

Lead-authored-by: Pavle Martinovic <34302662+Pajaraja@users.noreply.github.com>
Co-authored-by: pavle-martinovic_data <pavle.martinovic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>